requirements.txt
gradio
openai
google-generativeai
chromadb
sentence-transformers
pydantic
python-dotenv
